dev/requirements.txt
git+https://github.com/ubruhin/funq.git@504a5a9304b2354ffdfdff3fb89a7dba54fb38c3#egg=funq-server&subdirectory=server
git+https://github.com/ubruhin/funq.git@504a5a9304b2354ffdfdff3fb89a7dba54fb38c3#egg=funq&subdirectory=client
mss~=3.3.2
Pillow~=5.3.0
imageio~=2.4.1
